Reorder clean up code, so when USB dongle is removed, the application can be terminated without exceptions.
Don't know why (at least on wxMSW, not sure other platform):
When application is terminated normally (press X button
located on the right-top corner of frame or close function
in menu bar), the SimCubeFrame::~SimCubeFrame() will be
called before the SimCubeApp::OnExit(). But when we terminate
the application programmatically by wxApp::ExitMainLoop(),
the order will be reversed. This is a problem. So, put some
clean up in ~SimCubeApp(), it is always executed last.